chore: Handle rename
