+++ src/components/pages/homepage/gallery-items/chart-data-link.module.scss
