+++ src/components/pages/homepage/gallery-items/charts/chart.module.scss
