+++ src/components/pages/homepage/gallery-items/sidebar.module.scss
