+++ src/components/pages/homepage/gallery-items/title.js
