+++ src/components/pages/homepage/gallery-items/title.module.scss
