+++ src/components/pages/homepage/gallery-items/us-map/sidebar.module.scss
