+++ src/components/pages/homepage/gallery-items/us-map/states.js
