+++ src/components/pages/homepage/gallery-items/us-map/us-map.module.scss
