+++ src/components/pages/homepage/visualization-gallery.module.scss
