+++ src/pages/index.js
@@ -10,7 +10,7 @@ import BlogList from '~components/pages/homepage/blog-list'
 import Press from '~components/pages/homepage/press'
 import HomepageAPI from '~components/pages/homepage/api'
 import HomepageAbout from '~components/pages/homepage/about'
-import HomepageCarousel from '~components/pages/homepage/carousel'
+import HomepageVisualizationGallery from '~components/pages/homepage/visualization-gallery'
 import HomepageDatsets from '~components/pages/homepage/datasets'
 
 const Homepage = () => (
@@ -22,7 +22,7 @@ const Homepage = () => (
     <h1 className="a11y-only">The COVID Tracking Project</h1>
 
     <main id="main">
-      <HomepageCarousel />
+      <HomepageVisualizationGallery />
       <HomepageAbout />
       <HomepageDatsets />
